Added a fflush call and corrected the writeup in consequence.
The reason why I choose to change the code is that without any way to flush
output it happened to be *really* hard to recover the address of system() for
the exercise 8: when piped the process would segfault before flushing the
buffer and the target command would get nothing.

I found a way to bypass this problem on the original problem using a
hiddeous exploit that basically printed a lot of garbage after outputting the
address to fill the buffer and force flushing. However I do not expect
someone still learning to resort to such length.

The challenge is interesting though, if you are confident in your abilities
you may want to try beating the original program as it was.